+++ pkg/distributor/distributor.go
@@ -154,6 +154,7 @@ type Config struct {
 	ExtendWrites             bool   `yaml:"extend_writes"`
 	SignWriteRequestsEnabled bool   `yaml:"sign_write_requests"`
 	UseStreamPush            bool   `yaml:"use_stream_push"`
+	RemoteWriteV2Enabled     bool   `yaml:"remote_writev2_enabled"`
 
 	// Distributors ring
 	DistributorRing RingConfig `yaml:"ring"`
@@ -214,6 +215,7 @@ func (cfg *Config) RegisterFlags(f *flag.FlagSet) {
 	f.BoolVar(&cfg.ExtendWrites, "distributor.extend-writes", true, "Try writing to an additional ingester in the presence of an ingester not in the ACTIVE state. It is useful to disable this along with -ingester.unregister-on-shutdown=false in order to not spread samples to extra ingesters during rolling restarts with consistent naming.")
 	f.BoolVar(&cfg.ZoneResultsQuorumMetadata, "distributor.zone-results-quorum-metadata", false, "Experimental, this flag may change in the future. If zone awareness and this both enabled, when querying metadata APIs (labels names and values for now), only results from quorum number of zones will be included.")
 	f.IntVar(&cfg.NumPushWorkers, "distributor.num-push-workers", 0, "EXPERIMENTAL: Number of go routines to handle push calls from distributors to ingesters. When no workers are available, a new goroutine will be spawned automatically. If set to 0 (default), workers are disabled, and a new goroutine will be created for each push request.")
+	f.BoolVar(&cfg.RemoteWriteV2Enabled, "distributor.remote-writev2-enabled", false, "EXPERIMENTAL: If true, accept prometheus remote write v2 protocol push request.")
 
 	f.Float64Var(&cfg.InstanceLimits.MaxIngestionRate, "distributor.instance-limits.max-ingestion-rate", 0, "Max ingestion rate (samples/sec) that this distributor will accept. This limit is per-distributor, not per-tenant. Additional push requests will be rejected. Current ingestion rate is computed as exponentially weighted moving average, updated every second. 0 = unlimited.")
 	f.IntVar(&cfg.InstanceLimits.MaxInflightPushRequests, "distributor.instance-limits.max-inflight-push-requests", 0, "Max inflight push requests that this distributor can handle. This limit is per-distributor, not per-tenant. Additional requests will be rejected. 0 = unlimited.")
@@ -823,7 +825,17 @@ func (d *Distributor) Push(ctx context.Context, req *cortexpb.WriteRequest) (*co
 		return nil, err
 	}
 
-	return &cortexpb.WriteResponse{}, firstPartialErr
+	resp := &cortexpb.WriteResponse{}
+	if d.cfg.RemoteWriteV2Enabled {
+		// We simply expose validated samples, histograms, and exemplars
+		// to the header. We should improve it to expose the actual
+		// written values by the Ingesters.
+		resp.Samples = int64(validatedFloatSamples)
+		resp.Histograms = int64(validatedHistogramSamples)
+		resp.Exemplars = int64(validatedExemplars)
+	}
+
+	return resp, firstPartialErr
 }
 
 func (d *Distributor) updateLabelSetMetrics() {