+++ pkg/ingester/metrics_test.go
@@ -240,6 +240,18 @@ func TestTSDBMetrics(t *testing.T) {
 			# TYPE cortex_ingester_tsdb_wal_corruptions_total counter
 			cortex_ingester_tsdb_wal_corruptions_total 2.676537e+06
 
+			# HELP cortex_ingester_tsdb_wal_replay_unknown_refs_total Total number of unknown series references encountered during TSDB WAL replay.
+			# TYPE cortex_ingester_tsdb_wal_replay_unknown_refs_total counter
+			cortex_ingester_tsdb_wal_replay_unknown_refs_total{type="series"} 300
+			cortex_ingester_tsdb_wal_replay_unknown_refs_total{type="samples"} 303
+			cortex_ingester_tsdb_wal_replay_unknown_refs_total{type="metadata"} 306
+
+			# HELP cortex_ingester_tsdb_wbl_replay_unknown_refs_total Total number of unknown series references encountered during TSDB WBL replay.
+			# TYPE cortex_ingester_tsdb_wbl_replay_unknown_refs_total counter
+			cortex_ingester_tsdb_wbl_replay_unknown_refs_total{type="exemplars"} 300
+			cortex_ingester_tsdb_wbl_replay_unknown_refs_total{type="histograms"} 303
+			cortex_ingester_tsdb_wbl_replay_unknown_refs_total{type="tombstones"} 306
+
 			# HELP cortex_ingester_tsdb_wal_writes_failed_total Total number of TSDB WAL writes that failed.
 			# TYPE cortex_ingester_tsdb_wal_writes_failed_total counter
 			cortex_ingester_tsdb_wal_writes_failed_total 1486965
@@ -504,6 +516,18 @@ func TestTSDBMetricsWithRemoval(t *testing.T) {
 			# HELP cortex_ingester_tsdb_wal_corruptions_total Total number of TSDB WAL corruptions.
 			# TYPE cortex_ingester_tsdb_wal_corruptions_total counter
 			cortex_ingester_tsdb_wal_corruptions_total 2.676537e+06
+
+			# HELP cortex_ingester_tsdb_wal_replay_unknown_refs_total Total number of unknown series references encountered during TSDB WAL replay.
+			# TYPE cortex_ingester_tsdb_wal_replay_unknown_refs_total counter
+			cortex_ingester_tsdb_wal_replay_unknown_refs_total{type="series"} 300
+			cortex_ingester_tsdb_wal_replay_unknown_refs_total{type="samples"} 303
+			cortex_ingester_tsdb_wal_replay_unknown_refs_total{type="metadata"} 306
+
+			# HELP cortex_ingester_tsdb_wbl_replay_unknown_refs_total Total number of unknown series references encountered during TSDB WBL replay.
+			# TYPE cortex_ingester_tsdb_wbl_replay_unknown_refs_total counter
+			cortex_ingester_tsdb_wbl_replay_unknown_refs_total{type="exemplars"} 300
+			cortex_ingester_tsdb_wbl_replay_unknown_refs_total{type="histograms"} 303
+			cortex_ingester_tsdb_wbl_replay_unknown_refs_total{type="tombstones"} 306
 
 			# HELP cortex_ingester_tsdb_wal_writes_failed_total Total number of TSDB WAL writes that failed.
 			# TYPE cortex_ingester_tsdb_wal_writes_failed_total counter
@@ -882,6 +906,22 @@ func populateTSDBMetrics(base float64) *prometheus.Registry {
 		Help: "Total number snapshot replays that failed.",
 	})
 	snapshotReplayErrorTotal.Add(103)
+
+	walReplayUnknownRefsTotal := promauto.With(r).NewCounterVec(prometheus.CounterOpts{
+		Name: "prometheus_tsdb_wal_replay_unknown_refs_total",
+		Help: "Total number of unknown series references encountered during WAL replay.",
+	}, []string{"type"})
+	walReplayUnknownRefsTotal.WithLabelValues(typeSeries).Add(100)
+	walReplayUnknownRefsTotal.WithLabelValues(typeSamples).Add(101)
+	walReplayUnknownRefsTotal.WithLabelValues(typeMetadata).Add(102)
+
+	wblReplayUnknownRefsTotal := promauto.With(r).NewCounterVec(prometheus.CounterOpts{
+		Name: "prometheus_tsdb_wbl_replay_unknown_refs_total",
+		Help: "Total number of unknown series references encountered during WBL replay.",
+	}, []string{"type"})
+	wblReplayUnknownRefsTotal.WithLabelValues(typeExemplars).Add(100)
+	wblReplayUnknownRefsTotal.WithLabelValues(typeHistograms).Add(101)
+	wblReplayUnknownRefsTotal.WithLabelValues(typeTombstones).Add(102)
 
 	oooHistogram := promauto.With(r).NewHistogram(prometheus.HistogramOpts{
 		Name: "prometheus_tsdb_sample_ooo_delta",